python-fastapi-asyncpg/pyproject.toml
[project]
name = "python-fastapi-asyncpg"
version = "0.1.0"
description = "python-fastapi-asyncpg template"
requires-python = ">=3.13"

dependencies = [
    "fastapi",
    "uvicorn",
    "python-json-logger",
    "asyncpg",
    "prometheus-fastapi-instrumentator",
    "alembic",
    "psycopg2-binary",
    "httpx",
    "pydantic-settings",
]

[project.optional-dependencies]
dev = [
    "black",
    "mypy",
    "flake8",
    "pytest",
    "pytest-mock",
    "pytest-asyncio",
    "isort",
]

[tool.black]
line-length = 100

[tool.isort]
profile="black"
line_length = 100

[tool.mypy]
strict = true
ignore_missing_imports = true
